+++ spec/lcm/integration/data_helper_spec.rb
@@ -109,11 +109,50 @@ basic_params_without_schema = {
   }
 }
 
+snowflake_basic_params = {
+  "snowflake_client"=> {
+    "connection"=> {
+      "url"=> "jdbc:snowflake://gooddata.snowflakecomputing.com",
+      "authentication"=> {
+        "basic"=> {
+          "userName"=> "cornflake",
+          "password"=> ConnectionHelper::SECRETS[:redshift_password]
+        }
+      },
+      "database"=> "PHONG_DEV",
+      "warehouse"=> "PHONG_DEV_DWH"
+    }
+  },
+  "input_source"=> {
+    "type"=> "snowflake",
+    "query"=> "SELECT * FROM customers where CP__CUSTKEY='cus3'"
+  }
+}
+
+bigquery_basic_params = {
+  "bigquery_client"=> {
+    "connection"=> {
+      "authentication"=> {
+        "serviceAccount"=> {
+          "clientEmail"=> "gdc-bigquery-pipe@gdc-us-dev.iam.gserviceaccount.com",
+          }
+      },
+      "project"=> "gdc-us-dev",
+      "schema"=> "lcm_test"
+    }
+  },
+  "input_source"=> {
+    "type"=> "bigquery",
+    "query"=> "select * from employees order by EmployeeID"
+  }
+}
+
 describe 'data helper', :vcr do
 
   it 'connect to redshift with IAM authentication' do
     data_helper = GoodData::Helpers::DataSource.new(iam_params['input_source'])
     file_path = data_helper.realize(iam_params)
+    puts "redshift iam: #{file_path}"
     data = File.open('spec/data/redshift_data2.csv').read
     expect(data).to eq File.open(file_path).read
   end
@@ -121,6 +160,7 @@ describe 'data helper', :vcr do
   it 'connect to redshift with BASIC authentication' do
     data_helper = GoodData::Helpers::DataSource.new(basic_params['input_source'])
     file_path = data_helper.realize(basic_params)
+    puts "redshift basic 1: #{file_path}"
     data = File.open('spec/data/redshift_data2.csv').read
     expect(data).to eq File.open(file_path).read
   end
@@ -128,6 +168,7 @@ describe 'data helper', :vcr do
   it 'connect to redshift with BASIC authentication without schema' do
     data_helper = GoodData::Helpers::DataSource.new(basic_params_without_schema['input_source'])
     file_path = data_helper.realize(basic_params_without_schema)
+    puts "redshift basic 2: #{file_path}"
     data = File.open('spec/data/redshift_data.csv').read
     expect(data).to eq File.open(file_path).read
   end
@@ -135,6 +176,7 @@ describe 'data helper', :vcr do
   it 'connect to redshift with BASIC authentication and dynamic source' do
     data_helper = GoodData::Helpers::DataSource.new(basic_params_dynamic_source['dynamic_params']['input_source'])
     file_path = data_helper.realize(basic_params_dynamic_source)
+    puts "redshift basic 3: #{file_path}"
     data = File.open('spec/data/redshift_data2.csv').read
     expect(data).to eq File.open(file_path).read
   end
@@ -142,7 +184,29 @@ describe 'data helper', :vcr do
   it 'connect to redshift with BASIC authentication and url has parameter' do
     data_helper = GoodData::Helpers::DataSource.new(basic_params_url_parameters['input_source'])
     file_path = data_helper.realize(basic_params_url_parameters)
+    puts "redshift basic 4: #{file_path}"
     data = File.open('spec/data/redshift_data2.csv').read
+    expect(data).to eq File.open(file_path).read
+  end
+
+  it 'connect to snowflake with BASIC authentication' do
+    data_helper = GoodData::Helpers::DataSource.new(snowflake_basic_params['input_source'])
+    file_path = data_helper.realize(snowflake_basic_params)
+    puts "snowflake: #{file_path}"
+    data = File.open('spec/data/snowflake_data.csv').read
+    expect(data).to eq File.open(file_path).read
+  end
+
+  it 'connect to bigquery with BASIC authentication' do
+    encryption_key = ENV['GD_SPEC_PASSWORD'] || ENV['BIA_ENCRYPTION_KEY']
+    bigquery_secret = File.open('spec/environment/bigquery_encrypted').read
+    decrypted = GoodData::Helpers.decrypt(bigquery_secret, encryption_key)
+    bigquery_basic_params['bigquery_client']['connection']['authentication']['serviceAccount']['privateKey'] = decrypted
+
+    data_helper = GoodData::Helpers::DataSource.new(bigquery_basic_params['input_source'])
+    file_path = data_helper.realize(bigquery_basic_params)
+    puts "bigquery: #{file_path}"
+    data = File.open('spec/data/bigquery_data.csv').read
     expect(data).to eq File.open(file_path).read
   end
 end